[fstools]Do not export any ubi/mtd headers/libraries to staging dir (HOP-2321)
fstools exports a bunch of ubi header files and libubi-utils.a. This was done for a legacy
reasons and is now not needed anymore and actually interfering with
mtd-utils when ubi headers are needed because the contents of libubi.h
are different.

Issue: HOP-2321 SWUpdate: Update fails when using ubivol handler
GitOrigin-RevId: 95107267717d394aaede9a084279300134dd696c
(cherry picked from commit b8af6c77b4ee46a9204d03e8e4f26248213a8042)