[fstools]Do not export any ubi/mtd headers/libraries to staging dir (HOP-2321)
fstools exports a bunch of ubi header files and libubi-utils.a. This was done for a legacy
reasons and is now not needed anymore and actually interfering with
mtd-utils when ubi headers are needed because the contents of libubi.h
are different.

Issue: HOP-2321 SWUpdate: Update fails when using ubivol handler
GitOrigin-RevId: ff72c69f18b840d000037e0b0a311dbf1d4cd340
(cherry picked from commit 1d731d4744c20c5a585f22c12204680731dc1a60)