[fstools]Do not export any ubi/mtd headers/libraries to staging dir (HOP-2321)
fstools exports a bunch of ubi header files and libubi-utils.a. This was done for a legacy
reasons and is now not needed anymore and actually interfering with
mtd-utils when ubi headers are needed because the contents of libubi.h
are different.

Issue: HOP-2321 SWUpdate: Update fails when using ubivol handler
GitOrigin-RevId: a3622c8f8b4d1cfadfac3f946a3a85647b59a869
(cherry picked from commit a43510826d0a79f2de304a854e84e55ebbb3f085)