tr181-device: Add custom mapping file
The tr181-device plugin is responsible for the Device. mapping in the
datamodel. The file used in this plugin is static and the list contains
more than actually available in prplOS. This commit overrides that
static odl file and sets a file that is in line with what is availble in
prplOS.

Signed-off-by: Matthias FRANCK <matthias.franck_ext@softathome.com>
(cherry picked from commit ff65c2cad4ce77f27d7a0fc8c745267680d60c20)