tr181-device: Add custom mapping file
The tr181-device plugin is responsible for the Device. mapping in the
datamodel. The file used in this plugin is static and the list contains
more than actually available in prplOS. This commit overrides that
static odl file and sets a file that is in line with what is availble in
prplOS.

Signed-off-by: Matthias FRANCK <matthias.franck_ext@softathome.com>
(cherry picked from commit 882d34cfc6a156799cb93d8834499fa1289f7112)
(cherry picked from commit 3b4fccead2901d5dbdc82d907390a490c6229782)