PROVISORY: ssh-server: enforce IPv4 as default config bug workaround
The default IPVersion has changed to 0 (= 4 and 6). Due to a bug in
ssh-server IPv4 is not working if not all parameters are set for IPv6 as
well when creating a new SSH server instance. This commits keeps te
previous setting and sets the IP version back to v4.  This commit can be
reverted as soon as this bug is solved.

References: PCF-1086
Signed-off-by: Matthias FRANCK <matthias.franck_ext@softathome.com>
(cherry picked from commit 9f275147a9180ecbe493e67e8337281b47fe0483)