WIP: mod-amxb-usp: workaround poll timeout by increasing it to 3 seconds
strace we took on boot showed that the handshake does complete, but
after the 2 seconds deadline. As shared during prplOs increasing the
timeout here won't be a long term solution as, while waiting on that
poll, tr181-device won't proxy anymore  other request on regulare ubus
on Device. which is very central in the overall system.  Redoing the
handshake itself would fail due to duplicate endpoint error sent by
obuspa.  So the handshake/register flow will need to be reworked to
account for bigger delay without blocking  and allow retry of the
register part.

As it will likely take  a couple of weeks to rework the call flow and
properly test it, we might just temporarily increase the poll timeout to
3 or 4 seconds for immediate purpose of unlocking the tests.

So lets start with the 3s timeout.

Suggested-by: Vincent Harle <vincent.harle@softathome.com>
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>
(cherry picked from commit f7e1e8e58e09c78af8c423273512e9ff5c790b08)