[odhcp6c] add patch to prevent sending of router sollicitations
According to TR181, setting the RSEnable parameter to false should result in no router solicitation messages to be sent from the interface. This seems not the case, the messages are always seen regardless of the parameter value.
To solve this patch odhcp6c to not automatically sends Router Sollicitations as well.

This patch should be properly reworked to make it upstreamable. This progress is tracked in PCF-950

GitOrigin-RevId: 79d73313339530291cc2b659b5e77a1afbf381b2
(cherry picked from commit 10ead25e84ab22519b669fdaf21f7d5fea6831d1)