[odhcp6c] add patch to prevent sending of router sollicitations
According to TR181, setting the RSEnable parameter to false should result in no router solicitation messages to be sent from the interface. This seems not the case, the messages are always seen regardless of the parameter value.
To solve this patch odhcp6c to not automatically sends Router Sollicitations as well.

This patch should be properly reworked to make it upstreamable. This progress is tracked in PCF-950

GitOrigin-RevId: df618819bc948b5dab2ba028548c749d4ed4533e
(cherry picked from commit 94873acd1cbd8e3fbfc7f8c1fc787c47fa02bc62)