Issue: HOP-5738 [odhcp6c] Wrong retransmission of DHCPv6 Solicit [fix]
When the DHCPv6 client sends a DHCPv6 Solicit with both IA_NA and IA_PD options, and if the server replies with a status code = NoAddrsAvailable in the IA_NA option, then currently the DHCPv6 client sends a new Solicit with only the IA_PD option despite the fact that a prefix was sent by the server in the previous Advertise.

This behavior is described in https://datatracker.ietf.org/doc/html/rfc7550#section-4.2

The client must handle the case of a server that does not offer both valid IA_NA and IA_PD options when both are requested, according to RFC 7550. It should not send a new Solicit, but a Request. The client should however ignore the Advertise message if none of the IA_NA and IA_PD options are offered by the server.

GitOrigin-RevId: f90d3b3210f9f8189461a234a31b465f63073593
(cherry picked from commit 0327a7df1f156445070ef7ca849043de221b4243)