netdev-plugin: dm_link: create_transaction_from_link: fix wrong null check
GCC 12.3.0 rightfully complains about following:

      dm_link.c: In function 'create_transaction_from_link':
      dm_link.c:313:22: error: the comparison will always evaluate as 'true' for the address of 'name' will never be NULL [-Werror=address]
        313 |     if(((link->name) != NULL) && (*(link->name) != 0)) {
            |                      ^~
      In file included from dm_link.c:79:
      ../include_priv/dm_link.h:94:10: note: 'name' declared here
         94 |     char name[LINKNAMESPACE];
            |          ^~~~

This commit simplifies the condition checking whether the 'name' field
in the 'link' structure is non-empty. The original code erroneously
included a NULL check for 'name', which is not necessary and incorrect
because 'name' is an array, not a pointer, and thus cannot be NULL. The
corrected condition now directly checks if the first character of the
'name' array is not the null character, ensuring it's a non-empty
string. This change removes the compiler warning about the address
comparison that would always evaluate as true and makes the intent of
the check clearer.

References: PCF-1011
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>
(cherry picked from commit bd2a226e13ade7ab2b49b6a4384b237090c118f1)