[chrony] disable nts
When we converted chrony to chrony-prpl, this line should have been changed as well. Now the chrony-prpl package was building with nts support.
It turns out chrony has some mechanism to detect if gnutls is available in the environment. In this case, when nts support is enabled and when gnutls is available, the package is build with gnutls. If gnutls is not available, the package is simply build without gnutls support.

As a result, the chrony package is sometimes build with gnutls and sometimes without, depending on openwrt compilation order. By disabling nts support, we also disable the gnutls support.

GitOrigin-RevId: e75f84cc21eab2851784ee16ac46b1831d1710f8
(cherry picked from commit c38b69c70b6bb9422c21181ae9e3d54989d6bbd2)