[chrony] disable nts
When we converted chrony to chrony-prpl, this line should have been changed as well. Now the chrony-prpl package was building with nts support.
It turns out chrony has some mechanism to detect if gnutls is available in the environment. In this case, when nts support is enabled and when gnutls is available, the package is build with gnutls. If gnutls is not available, the package is simply build without gnutls support.

As a result, the chrony package is sometimes build with gnutls and sometimes without, depending on openwrt compilation order. By disabling nts support, we also disable the gnutls support.

GitOrigin-RevId: 9ab705f59351a84e2058c7aee6276591a0bc3e03
(cherry picked from commit 0f17f06dfa12ef0bf83e862e56bfa855ed0994f4)