treewide: add ap runlevel files
For HOP-7080 we need to be able to remove packages at runtime when we are in a certain runlevel.
Therefore we add runlevel scripts that get only excecuted when we are in that particular runlevel.

Issue: HOP-7080 add support for multiple runlevels
GitOrigin-RevId: fe4d688b1328d7cc6e70690e16af7ed9a8150965
(cherry picked from commit a7a5cb0ca53eb46612933a20bd6f1560580bbaba)