PROVISORY: prpl-configuration: add networklayout.json file for wnc-freedom
It looks like the order of WiFi.Radio.X interfaces is not deterministic. This is caused because the networklayout.json file that is generated has no fixed order of the radio interfaces.

As a workaround, a static networklayout.json is added for the wnc-freedom box.

This issue is related to PPW-20 and should probably be solved by integrating prpl-foundation/prplos/feeds/feed-qca@84bd94fa

Signed-off-by: Matthias FRANCK <matthias.franck@softathome.com>
(cherry picked from commit 912d86370121edf4a470ed87ce9387041f9edca7)