PROVISORY: prpl-configuration: dhcpv4-manager: set lcm dhcp pool to 192.168.3.0/24
In prpl-foundation/components/core/plugins/tr181-ip@51e1087b the odl files in IP manager are adapted to use ucode files.
These files now create IP configurations depending on the number of bridges in the networklayout.json (formerly we only had 3 static bridges).

As a result the ip bridge gets now 192.168.3.1 as ip adress, not longer 192.168.5.1.

The dhcpv4 server should be updated in the future as well, to create in a similar way dhcp configuration for every bridge. This is not yet done, and hence the dhcpv4 server is still giving ip adresses in 192.168.5.0/24 range.

This commits sets the dhcp pool for the lcm bridge to 192.168.3.0/24 for the time being.

Signed-off-by: Matthias FRANCK <matthias.franck@softathome.com>
(cherry picked from commit 78b363434f33953b061cec92b07c4887eacbc8da)