time-manager: upstep to v2.7.2
Time-manager sometimes takes to long to synchronize (+90s).
Time-manager tries to start the sync when the wan interface becomes up. 
If it takes a while between wan up and receiving a wan IP address the 
timeout has significantly increased causing the sync to take a long time.

Instead of listening for the wan to become up, we now listen on the wan 
receiving an addresses. Restarting the sync when a change to the 
addresses occurs.

With the upstep following changes are introduced:
-Issue: HOP-7379 [time-manager] Sync fails when in sequential mode
-Issue: HOP-7342 - Improve the way time-manager updates chrony daemon
-Issue: PCF-1168 - Time.Status takes more than 90secs to get 
   Synchronized state at startup
-Issue: PCF-1257 - time manager fails to synchronize time
-Issue: HOP-6945 PRPL/COVERITY: time-manager issue

Fixes: PCF-1168
Signed-off-by: Yüce Kürüm <yuce.kurum@mind.be>
Signed-off-by: Yüce Kürüm <yuce.kurum_ext@softathome.com>
(cherry picked from commit 5235b49bd40ba9d3814711e94e0e5e798de68adb)