[tr181-mqtt] Client unable to connect to Broker using PKCS#11 Query Key
A previous commit introduced a patch to support openssl3 providers in
mosquitto library. However, this patch was missing on a test made in a
high level API, making the API check if a file path exists, even when
it is an openssl3 provider URI.

This commit bypasses the check (it is irrelevant with providers) if we
are using openssl3 as SSL backend.

Issue: [tr181-mqtt] Client unable to connect to Broker using PKCS#11
Query Key - HOP-7742
GitOrigin-RevId: 0a72aa5c187cddec52aa425b1ccb4562021a2b18
(cherry picked from commit cf8f779270fd98ce0e0d92ab6b353c2f1741ae60)