[tr181-mqtt] Client unable to connect to Broker using PKCS#11 Query Key
A previous commit introduced a patch to support openssl3 providers in
mosquitto library. However, this patch was missing on a test made in a
high level API, making the API check if a file path exists, even when
it is an openssl3 provider URI.

This commit bypasses the check (it is irrelevant with providers) if we
are using openssl3 as SSL backend.

Issue: [tr181-mqtt] Client unable to connect to Broker using PKCS#11
Query Key - HOP-7742
GitOrigin-RevId: f562b8601ee876a2398f0921f3a42e4c08175c7a
(cherry picked from commit 0f1b012d7ae29cd6aac6c0c31961abc7508d9ec6)