p11-kit: patch to add build option for secure_getenv
p11-kit client relies on reading environment variables.

Access to environment variables will fail for processes that changed userid, groupid or capabilities. This is the case for plugins that have configurations to not execute as root.

To have this type of plugins use pkcs11 uris, we need to have some way to access the information set in some environment variables. This patch adds a build option to allow reading of the needed environment variables, it can be set to true to remove all access again.

Issue: HGWPR-33663 tr181-mqttbroker unable to read pkcs uri when not run as root
GitOrigin-RevId: 8cd574b138a60d410cbe088a266c1957c4326e5c
(cherry picked from commit 5a5561cecb1112f2135ab43597979864c8c0ee1c)