libqmi: import v1.34.0
This is a dependency of the modemmanager package.
We need this newer version, since we also imported a newer version of modemmanager (v1.22.0) that depends on this lib.

This package is imported from the openwrt packages feed at e828f98ba907840cb6418856c2163821946d2c87 (fwupd: add package)

References: PCF-1636
Signed-off-by: Matthias Franck <matthias.franck@softathome.com>