unbound-prpl: increase PKG_RELEASE
In fb13a6a652ad78258419d3ad3d0adfe579abc847 and faacfb6be3c917eac7c9d4f69b360565412c1da0 we added patches on top of unbound v1.22.0.
We however forgot to increase the PKG_RELEASE to make it clear that this is a patched version of unbound.

So let's set PKG_RELEASE to 2 to indicate this is a patched version so that this is visible from the sbom, opkg, ...

Signed-off-by: Matthias Franck <matthias.franck@softathome.com>