efr32-boot-config: move package from feed-qca
We've the same Silabs EFR32MG21 MCU on both Fast Track reference
platforms (OSPv2 and Freedom), being wired to the SoC in the similar way
(GPIO pins).

In the default configuration the nReset PIN on EFR32MG21 is being held
down by the SoC GPIO pin (usually low by default), so the EFR32MG21 is
being held in the reset state, thus making the MCU unavailable to the
system/application.

On Freedom we've efr32-boot-config package providing init script, which
puts EFR32MG21 MCU out of the reset state, disables the recovery
functionality and thus boot the MCU into application firmware.

On OSPv2 we're currently missing this functionality and the plan is to
reuse the common bits and share the package in the generic fashion.

As a starting point, we're currently moving the package from feed-qca as
it is, so the modifications for OSPv2 board can be done in seperate
commit later.

References: PCF-1297
Signed-off-by: Petr Štetiar <petr.stetiar@prplfoundation.org>