USP: Define the default USP backend capabilities. (PCF-1884)
Problem:
The default values of USP capabilities are 'resolve', 'has' and 'describe'.
Discovering data models via the ambiorix describe functionality is not supported when OBUSPA is used (because it requires calling the _describe method, which is not allowed by OBUSPA.)

Solution:
Set the capabilities to 'resolve', 'has' is sufficient and will avoid error logs.

Issue: Define the default USP backend capabilities. - PCF-1884
Signed-off-by: Adnen Masmoudi <adnen.masmoudi_ext@sofathome.com>