libp11: configure openssl pkcs11 engine
Add openssl config file that configures pkcs11 engine.
Enable it using openwrt standard configuration mechanism - uci.
Use uci-defaults method to append to /etc/config/openssl as the file is
already a part of libopenssl-conf package and can't be provided by two.

Additionaly modify the engine path from:
/usr/lib/engines
to:
/usr/lib/engines-3

As this is the location that openssl on openwrt uses.

References: PCF-2020
Signed-off-by: Piotr Nakraszewicz <piotr.nakraszewicz@consult.red>