softhsm2: adjust the init script to cram tests procedure
The dev certificates generated during the boot are found in cram test to
be created while time on the box is still 1970s. Since those are for
development/test purposes only, make them valid for a 100 years.

Also move the init script so that it executes just before tr181-security
as it is relatede the most to this module.

Signed-off-by: Piotr Nakraszewicz <piotr.nakraszewicz@consult.red>