prpl-configuration: tr181-dns:add missing DNS.Zone object to USP bus
Problem:
The Device.DNS.Zone. data model object was not visible on the USP bus.

Cause:
The DNS.Zone object was missing from the tr181-dns_usp.odl file, which defines the TR-181 objects exposed over USP.

Solution:
Added the Device.DNS.Zone. object to the list of USP registrations in tr181-dns_usp.odl to ensure it is properly mapped and exposed on the USP bus.

References: PPW-681

Signed-off-by: Sarra Kacem <sarra.kacem_ext@softathome.com>