libmbim: libqmi: remove imported packages
These newer versions of the libs were needed because we wanted to use a newer version of modemmanager in the past (v1.22.0).
Since we are now using openwrt-24.10, we already get newer versions of these libs that are compatible with modemmanager v1.22.0.

So let's cleanup these imports.

The modemmanager package cannot be removed, since we still need to update the dependency on the ppp package.

References: PPW-1098

Signed-off-by: Matthias Franck <matthias.franck@softathome.com>