tr181-firewall: fix installation of defaults
During integration of 42fac30be2b267898231482d65ee37672bc0e3ad CONFIG_GATEWAY and CONFIG_ACCESSPOINT variables are passed to the makefiles.
This leads to the fact that CONFIG_GATEWAY is undefined and as a result some defaults are not properly installed in the rootfs.

So let's fix it by not defining them as empty.

Signed-off-by: Matthias Franck <matthias.franck@softathome.com>